+++ Assets/Samples/RequestHandler/SampleMessage.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace Causeless3t.Network.Protocol
+{
+    [Serializable]
+    public class SampleReq : BaseRequest
+    {
+        public int SamepleParam;
+    }
+
+    [Serializable]
+    public class SampleAns : BaseResponse
+    {
+        
+    }
+}